pyproject.toml
[project]
name = "jobfit-server"
version = "0.1.0"
description = "JobFit Backend Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "elevenlabs>=1.0.0",
    "websockets>=13.1",
    "pydantic-settings>=2.12.0",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.22",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "numpy>=2.4.1",
    # Claude Agent + LangGraph
    "langgraph>=0.2.0",
    "anthropic>=0.40.0",
    # Git Push
    "PyGithub>=2.1.1",
    # Database (Replit PostgreSQL)
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    # STT (Deepgram)
    "deepgram-sdk>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server/app"]

[dependency-groups]
dev = [{include-group = "lint"}, {include-group = "test"}]
lint = [
    "ruff>=0.15.1",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

# --- Tool Configuration ---

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (security)
    "ASYNC",# flake8-async
]
ignore = [
    "E501",  # 줄 길이 — ruff format이 코드는 처리, 프롬프트 문자열은 예외
    "S101",  # assert 사용 허용 (테스트)
    "S104",  # 0.0.0.0 바인딩 허용 (Replit 배포 필수)
    "B008",  # Depends() in default args (FastAPI 표준 패턴)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "ASYNC230"]
"server/app/services/github_service.py" = ["S110", "S112"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=server/app",
    "--cov-report=term-missing",
    "--strict-markers",
]
